requirements.txt
apex==0.9.10dev
cityscapesscripts==2.2.0
einops==0.8.0
ftfy==6.2.0
matplotlib==3.9.0
mmdet==2.25.0
Pillow==10.3.0
prettytable==3.10.0
pycocotools==2.0.7
regex==2024.5.15
scipy==1.13.1
timm==1.0.3
opencv-python==4.8.1.78
xformers==0.0.22
mmcv-full==1.5.3
numpy==1.26.4
# torch==2.0.1
# torchaudio==2.0.2
# torchvision==0.15.2